split interferometer into modules / packages
