+++ autoarray/inversion/inversion/factory.py
@@ -2,7 +2,7 @@ from typing import Dict, List, Optional, Union
 
 from autoarray.dataset.imaging.imaging import Imaging
 from autoarray.dataset.imaging.w_tilde import WTildeImaging
-from autoarray.dataset.interferometer import WTildeInterferometer
+from autoarray.dataset.interferometer.w_tilde import WTildeInterferometer
 from autoarray.structures.arrays.uniform_2d import Array2D
 from autoarray.structures.visibilities import Visibilities
 from autoarray.structures.visibilities import VisibilitiesNoiseMap
