+++ autoarray/inversion/inversion/interferometer/w_tilde.py
@@ -6,7 +6,7 @@ from autoconf import cached_property
 from autoarray.inversion.inversion.interferometer.abstract import (
     AbstractInversionInterferometer,
 )
-from autoarray.dataset.interferometer import WTildeInterferometer
+from autoarray.dataset.interferometer.w_tilde import WTildeInterferometer
 from autoarray.inversion.linear_obj.linear_obj import LinearObj
 from autoarray.inversion.inversion.settings import SettingsInversion
 from autoarray.inversion.pixelization.mappers.abstract import AbstractMapper
