--- test_autoarray/dataset/files/interferometer/positions.dat
@@ -1,1 +1,0 @@
-[(1.0, 1.0), (2.0, 2.0)]
