+++ test_autoarray/dataset/interferometer/files/3_multiple_hdu.fits
