+++ test_autoarray/dataset/interferometer/files/3x2_fours.fits
