+++ test_autoarray/dataset/interferometer/files/3x2_multiple_hdu.fits
