+++ test_autoarray/dataset/interferometer/files/3x2_threes.fits
